# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"InvControl Pro API","version":"0.1.0"},"paths":{"/products":{"post":{"summary":"Create Product","operationId":"create_product_products_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/products/{product_id}":{"get":{"summary":"Get Product","operationId":"get_product_products__product_id__get","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"integer","title":"Product Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sales":{"post":{"summary":"Create Sale","operationId":"create_sale_sales_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Sale"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/products/{product_id}/stock":{"put":{"summary":"Update Stock","operationId":"update_stock_products__product_id__stock_put","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"integer","title":"Product Id"}},{"name":"stock","in":"query","required":true,"schema":{"type":"integer","title":"Stock"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Product":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"price":{"type":"number","title":"Price"},"stock":{"type":"integer","title":"Stock"}},"type":"object","required":["id","name","price","stock"],"title":"Product"},"Sale":{"properties":{"product_id":{"type":"integer","title":"Product Id"},"quantity":{"type":"integer","title":"Quantity"}},"type":"object","required":["product_id","quantity"],"title":"Sale"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}